+++ framework/db/pgsql/QueryBuilder.php
@@ -338,6 +338,10 @@ class QueryBuilder extends \yii\db\QueryBuilder
         if (empty($uniqueNames)) {
             return $insertSql;
         }
+        if ($updateNames === []) {
+            // there are no columns to update
+            $updateColumns = false;
+        }
 
         if ($updateColumns === false) {
             return "$insertSql ON CONFLICT DO NOTHING";
@@ -367,6 +371,10 @@ class QueryBuilder extends \yii\db\QueryBuilder
         list($uniqueNames, $insertNames, $updateNames) = $this->prepareUpsertColumns($table, $insertColumns, $updateColumns, $constraints);
         if (empty($uniqueNames)) {
             return $this->insert($table, $insertColumns, $params);
+        }
+        if ($updateNames === []) {
+            // there are no columns to update
+            $updateColumns = false;
         }
 
         /** @var Schema $schema */